Updated File Structure added IMAT Model Method
Update filestructure to reflect plans for project, and moved over files
from older consensus-essentiality project (diversity submodule). Created
a method for generating a model using IMAT, with a constraint added to
ensure the fluxes in the model match the optimal IMAT solution. Created
a method for sampling these models, however, the sampling methods don't
work with integer problems which the IMAT model currently is. This
method could be adjusted to create IMAT models in a different way (such
as the various model update methods found in the diversity submodule),
but as of now does not work.

Additionally, tests were added for the updated IMAT submodule (including
tests that fail due to the above mentioned issue with sampling).

Added data files for tests, including small toy metabolic models and
enumeration results.